+++ lib/cobra_commander/component_tree.rb
@@ -140,11 +140,11 @@ module CobraCommander
         @deps ||= begin
           return [] unless node?
           json = JSON.parse(File.read(package_json_path))
-          format combined_deps(json)
+          combined_deps(json)
         end
       end
 
-      def format(deps)
+      def format_dependencies(deps)
         return [] if deps.nil?
         linked_deps = deps.select { |_, v| v.start_with? "link:" }
         linked_deps.map do |_, v|
@@ -163,7 +163,32 @@ module CobraCommander
       end
 
       def combined_deps(json)
-        Hash(json["dependencies"]).merge(Hash(json["devDependencies"]))
+        worskpace_dependencies = build_workspaces(json["workspaces"])
+        dependencies = format_dependencies Hash(json["dependencies"]).merge(Hash(json["devDependencies"]))
+        (dependencies + worskpace_dependencies).uniq
+      end
+
+      def build_workspaces(workspaces)
+        return [] if workspaces.nil?
+        workspaces.map do |workspace|
+          glob = "#{@root_path}/#{workspace}/package.json"
+          workspace_dependencies = Dir.glob(glob)
+          workspace_dependencies.map do |wd|
+            { name: component_name(wd), path: component_path(wd) }
+          end
+        end.flatten
+      end
+
+    private
+
+      def component_name(dir)
+        component_path(dir).split("/")[-1]
+      end
+
+      def component_path(dir)
+        return dir.split("/package.json")[0] if @root_path == "."
+
+        dir.split(@root_path)[-1].split("/package.json")[0]
       end
     end
   end