+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+* Tracks package.json `workspaces` in addition to `dependencies` and `devDependencies`. PR [#31](https://github.com/powerhome/cobra_commander/pull/31)
+
 ## Version 0.5.1 - 2018-10-15
 
 * Fix a bug with dependencies_of where it wouldn't match components further down the list of umbrella's dependencies